deep fake/deep fake/requirements.txt
Flask
torch
torchvision
timm
opencv-python
numpy
Pillow
werkzeug
